+++ examples/userguide/a_npc_arrays_triv.py
@@ -12,5 +12,5 @@ print("|v> =", v.to_ndarray())
 M_v = npc.tensordot(M, v, axes=[1, 0])
 print("M|v> =", M_v.to_ndarray())
 # M|v> = [ 4.+1.j  3.+0.j]
-print("<v|M|v> =", npc.inner(v.conj(), M_v))
+print("<v|M|v> =", npc.inner(v.conj(), M_v, axes='range'))
 # <v|M|v> = (24+0j)